backend/requirements.txt
# 📁 backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
Werkzeug
python-dotenv
psycopg2-binary